+++ packages/moengage_flutter/moengage_flutter_android/CHANGELOG.md
@@ -2,10 +2,10 @@
           
 # Next Release Date
 
-## Next Vesion
+## Next Version
 - Add `jsonEncode()` for missing APIs 
-- `moe-android-sdk` version updated to `13.05.01`
-- `inapp` version updated to `8.7.0`
+- `moe-android-sdk` version updated to `13.05.02`
+- `inapp` version updated to `8.7.1`
 
 # 25-11-2024