+++ packages/moengage_flutter/moengage_flutter_android/CHANGELOG.md
@@ -1,8 +1,8 @@
 # MoEngage Flutter Android Plugin
           
-# Next Release Date
+# 29-01-2025
 
-## Next Version
+## 3.2.0
 - Add `jsonEncode()` for missing APIs 
 - `moe-android-sdk` version updated to `13.05.02`
 - `inapp` version updated to `8.7.1`